1406: Cleanup code
Remove unused api_clients configuration from the DevelopmentTrustedDirectoryConfig

General cleanup based on IntelliJ code analysis.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1406
