+++ src/main/kotlin/com/forgerock/sapi/gateway/framework/configuration/ConfigurationManager.kt
@@ -27,7 +27,7 @@ annotation class ApisUnderTest(
     val apisUnderTest: Array<String> = []
 )
 
-class ConfigurationManager() :
+class ConfigurationManager :
     BeforeAllCallback {
 
     companion object Loader {
