+++ src/main/kotlin/com/forgerock/sapi/gateway/framework/configuration/DevelopmentTrustedDirectoryConfig.kt
@@ -2,7 +2,6 @@ package com.forgerock.sapi.gateway.framework.configuration
 
 data class DevelopmentTrustedDirectoryConfig(
     val name: String,
-    val apiClients: List<ApiClientConfig>,
     var getKeysUrl: String,
     var getTransportPemsUrl: String,
     var getSigningPemsUrl: String,
